Agent File/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.26.0
python-dotenv==1.0.0

# AI - Basic setup (LangChain will be added in Phase 2)
google-generativeai==0.3.2

# Database vector support
pgvector==0.2.4

# Scheduling
apscheduler==3.10.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
